+++ app/src/main/java/com/paw/key/domain/model/entity/archivedlist/ArchivedListEntity.kt
@@ -5,7 +5,7 @@ data class ArchivedListPostsEntity(
 )
 
 data class ArchivedListEntity(
-    val postId: Long,
+    val postId: Int,
     val createdAt: String,
     val isLiked: Boolean,
     val title: String,
